doing FreeBSD sysctl(3) with Rust
there are two approaches to use here: the first, and probably the
easiest, is the one I'm using here, with `sysctlbyname`, which takes a
CString and a bunch of pointers and spews back sysctl data.

the approach I'd like to use is something like

    let oid: [i32; 8] = [libc::CTL_HW, libc::HW_NCPU];
    let mut t: [i32; 1] = [0];
    let mut tlen: size_t = 4;

    unsafe {
        libc::sysctl(oid.as_ptr(), 2 as size_t,
	    mem::transmute(&mut t), &mut tlen,
	    ptr::null(), 0) }

this didn't work, though, probably for reasons that'll become obvious
the moment I try doing something else.

stuff I haven't tried yet:

- setting sysctl(3) values (doesn't really make sense in this context)
- more complex structures

- the Staff OIDs (0.x.*) for types, but that starts to smell like
  runtime type analysis and in any case the Staff OIDs are
  ~undocumented~ kernel interfaces

woo hacks